Backend/requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.5
django-cors-headers==4.7.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
future==1.0.0
idna==3.10
invoke==2.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==7.9.0
oauthlib==3.3.1
paramiko==4.0.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyserial==3.5
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scp==0.15.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
textfsm==1.1.3
urllib3==2.5.0
